rename internal map
