+++ src/main/java/com/fasterxml/jackson/databind/util/internal/LinkedDeque.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fasterxml.jackson.databind.util.clhm;
+package com.fasterxml.jackson.databind.util.internal;
 
 import java.util.AbstractCollection;
 import java.util.Collection;
