+++ src/main/java/com/fasterxml/jackson/databind/util/internal/Weighers.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fasterxml.jackson.databind.util.clhm;
+package com.fasterxml.jackson.databind.util.internal;
 
-import static com.fasterxml.jackson.databind.util.clhm.ConcurrentLinkedHashMap.checkNotNull;
+import static com.fasterxml.jackson.databind.util.internal.PrivateMaxEntriesMap.checkNotNull;
 
 import java.io.Serializable;
