+++ src/main/java/com/fasterxml/jackson/databind/util/internal/package-info.java
@@ -22,19 +22,19 @@
  * missing features compared to the full <a href="http://code.google.com/p/concurrentlinkedhashmap/">
  * http://code.google.com/p/concurrentlinkedhashmap/</a> implementation.
  * <p>
- * {@link com.fasterxml.jackson.databind.util.clhm.Weigher} is a simple interface
+ * {@link com.fasterxml.jackson.databind.util.internal.Weigher} is a simple interface
  * for determining how many units of capacity an entry consumes. Depending on
  * which concrete Weigher class is used, an entry may consume a different amount
  * of space within the cache. The
- * {@link com.fasterxml.jackson.databind.util.clhm.Weighers} class provides
+ * {@link com.fasterxml.jackson.databind.util.internal.Weighers} class provides
  * utility methods for obtaining the most common kinds of implementations.
  * <p>
- * {@link com.fasterxml.jackson.databind.util.clhm.EvictionListener} provides the
+ * {@link com.fasterxml.jackson.databind.util.internal.EvictionListener} provides the
  * ability to be notified when an entry is evicted from the map. An eviction
  * occurs when the entry was automatically removed due to the map exceeding a
  * capacity threshold. It is not called when an entry was explicitly removed.
  * <p>
- * The {@link com.fasterxml.jackson.databind.util.clhm.ConcurrentLinkedHashMap}
+ * The {@link com.fasterxml.jackson.databind.util.internal.PrivateMaxEntriesMap}
  * class supplies an efficient, scalable, thread-safe, bounded map. As with the
  * <tt>Java Collections Framework</tt> the "Concurrent" prefix is used to
  * indicate that the map is not governed by a single exclusion lock.
@@ -42,4 +42,4 @@
  * @see <a href="http://code.google.com/p/concurrentlinkedhashmap/">
  *      http://code.google.com/p/concurrentlinkedhashmap/</a>
  */
-package com.fasterxml.jackson.databind.util.clhm;
+package com.fasterxml.jackson.databind.util.internal;
