+++ src/test/java/com/fasterxml/jackson/databind/util/internal/MapTestFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fasterxml.jackson.databind.util.clhm;
+package com.fasterxml.jackson.databind.util.internal;
 
 import java.util.Map;
 import java.util.function.Supplier;
